项目优化，解决Moment item图片数目显示错乱问题，添加注销登录等

+++ app/src/main/java/org/scau/mimi/gson/LoginInfo.java
@@ -18,9 +18,9 @@ public class LoginInfo {
 
         public User user;
 
-        public int tmCreated;
+        public long tmCreated;
 
-        public int tmExpire;
+        public long tmExpire;
 
         public class User {
             public int uid;
@@ -29,7 +29,7 @@ public class LoginInfo {
 
             public String nname;
 
-            public int tmCreated;
+            public long tmCreated;
 
         }
